Addresses #13 - REST API documentation via OpenAPI
Addition of OpenAPI 3.1 Yaml file describing OIDPlus API.

This initial file has a considerable amount of placeholder data that should be updated, such as:
Descriptions.
Contact Info.
URLs.

It ommits some data, such as security information.

Some of it is speculative such as: The path variables. The response properties are in an array called 'Response'.

Someone more familiar with the API should check it thoroughly, confirm everything, and make updates where needed.

There is also some repetition in the data, which could be fixed with  but that should wait until it has been checked.